Add support for tokens in PEG parser
Add flags to the PEG parsing routine and a to indicate token
generation.  Token generation works as follow.  If the token generation
flag is set, then all unresolved identifiers whose name string is all
capital letters is given a definition node ID number in [-254, -128]
starting with -128 and going down.  In order for the cpeg parser to
match a this identifier, it must match a literal positive value in [128,
254].

The pegcc supports this functionality by adding a -t option to enable
the PML_GEN_TOKEN flag.  It also produces #define directives for each
token so accompanying C code can produce the correct tokens in their
relevant getc-function.  The names of the #define tokens default to
starting with CPEGTOK_, but pegcc allows one to override this with the
-T option.

This patch also adds a testtok.peg to test this functionality.

Also, while at it, remove end-of-line whitespace in PEG files and fix
calc10.peg to properly left-associate subtraction and division.